refactor: #13 - dataProcessor -> domain
