--- backend/ezgg/src/main/java/com/matching/ezgg/dataProcessor/riotApi/dto/PuuidDto.java
@@ -1,10 +1,0 @@
-package com.matching.ezgg.dataProcessor.riotApi.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class PuuidDto {
-	private String puuid;
-}
