+++ backend/ezgg/src/main/java/com/matching/ezgg/domain/memberInfo/util/MatchIdConvert.java
@@ -1,0 +1,35 @@
+package com.matching.ezgg.domain.memberInfo.util;
+
+import java.io.IOException;
+import java.util.List;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.core.type.TypeReference;
+import com.fasterxml.jackson.databind.ObjectMapper;
+
+import jakarta.persistence.AttributeConverter;
+import jakarta.persistence.Converter;
+
+@Converter
+public class MatchIdConvert implements AttributeConverter<List<String>, String> {
+
+	private final ObjectMapper objectMapper = new ObjectMapper();
+
+	@Override
+	public String convertToDatabaseColumn(List<String> attribute) {
+		try {
+			return objectMapper.writeValueAsString(attribute);
+		} catch (JsonProcessingException e) {
+			throw new IllegalArgumentException("List<String> → JSON 변환 실패", e);
+		}
+	}
+
+	@Override
+	public List<String> convertToEntityAttribute(String dbData) {
+		try {
+			return objectMapper.readValue(dbData, new TypeReference<List<String>>() {});
+		} catch (IOException e) {
+			throw new IllegalArgumentException("JSON → List<String> 변환 실패", e);
+		}
+	}
+}
