+++ backend/ezgg/src/main/java/com/matching/ezgg/domain/riotApi/dto/PuuidDto.java
@@ -1,0 +1,10 @@
+package com.matching.ezgg.domain.riotApi.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class PuuidDto {
+	private String puuid;
+}
