+++ backend/ezgg/src/main/java/com/matching/ezgg/member/controller/MemberController.java
@@ -11,9 +11,9 @@ import com.matching.ezgg.matching.dto.MemberDataBundleDto;
 import com.matching.ezgg.member.dto.SignupRequest;
 import com.matching.ezgg.member.dto.SignupResponse;
 import com.matching.ezgg.member.service.MemberService;
-import com.matching.ezgg.dataProcessor.memberInfo.entity.MemberInfo;
+import com.matching.ezgg.domain.memberInfo.entity.MemberInfo;
 import com.matching.ezgg.matching.service.MemberDataBundleService;
-import com.matching.ezgg.dataProcessor.memberInfo.service.MemberInfoService;
+import com.matching.ezgg.domain.memberInfo.service.MemberInfoService;
 import com.matching.ezgg.global.annotation.LoginUser;
 import com.matching.ezgg.member.jwt.filter.JWTUtil;
 import com.matching.ezgg.member.jwt.repository.RedisRefreshTokenRepository;
