+++ backend/ezgg/src/main/java/com/matching/ezgg/member/service/MemberService.java
@@ -4,15 +4,15 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.matching.ezgg.dataProcessor.riotApi.service.ApiService;
+import com.matching.ezgg.domain.riotApi.service.ApiService;
 import com.matching.ezgg.matching.service.MatchingService;
 import com.matching.ezgg.member.dto.SignupRequest;
 import com.matching.ezgg.member.dto.SignupResponse;
 import com.matching.ezgg.member.entity.Member;
 import com.matching.ezgg.member.repository.MemberRepository;
-import com.matching.ezgg.dataProcessor.memberInfo.entity.MemberInfo;
-import com.matching.ezgg.dataProcessor.memberInfo.repository.MemberInfoRepository;
-import com.matching.ezgg.dataProcessor.memberInfo.service.MemberInfoService;
+import com.matching.ezgg.domain.memberInfo.entity.MemberInfo;
+import com.matching.ezgg.domain.memberInfo.repository.MemberInfoRepository;
+import com.matching.ezgg.domain.memberInfo.service.MemberInfoService;
 import com.matching.ezgg.global.exception.ExistEmailException;
 import com.matching.ezgg.global.exception.ExistMemberIdException;
 import com.matching.ezgg.global.exception.ExistRiotUsernamException;
